Fix ambiguous Arduino String overload in Meanwell CAN metric publishing
Agent-Logs-Url: https://github.com/Jones14021/OpenDTU/sessions/22b24c7b-337d-43dd-8632-ff3eeeeb8ec9

+++ src/MeanwellCan.cpp
@@ -361,7 +361,7 @@ void MeanwellCanClass::decodeMeanwellPbn(const CanFrame& frame)
 
 void MeanwellCanClass::publishMetric(const String& topic, float value, const uint8_t decimals)
 {
-    MqttSettings.publish(topic, String(value, decimals));
+    MqttSettings.publish(topic, String(value, static_cast<unsigned int>(decimals)));
 }
 
 uint8_t MeanwellCanClass::readRegister(uint8_t address)
